Refactor: Sort out BlazorServer and Core dependencies. Move Condifurator code under Core project. BlazorServer clean up usings. Remove NetworkedAddonDataProvider and IAddonDataProvider.
