+++ BlazorServer/Pages/BagItemComponent.razor
@@ -1,8 +1,4 @@
-﻿@using SharedLib
-@using Core
-@using System.Xml
-
-<a href="@($"{WApi.ItemId}{BagItem.ItemId}")" target="_blank" data-wowhead="item=@BagItem.ItemId" class="small" style="text-decoration:none">
+﻿<a href="@($"{WApi.ItemId}{BagItem.ItemId}")" target="_blank" data-wowhead="item=@BagItem.ItemId" class="small" style="text-decoration:none">
     <b>
         <span style="color:@BagItemColour(BagItem.Item);font-weight:normal">
             @if (BagItem.Count > 1)
