+++ BlazorServer/Pages/BotHeader.razor
@@ -1,6 +1,6 @@
-﻿@using Core
-@inject Core.IAddonReader addonReader
-@inject Core.IBotController botController
+﻿@inject IAddonReader addonReader
+@inject IBotController botController
+
 @implements IDisposable
 
 <div class="card @(Hide ? "hide" : string.Empty)">
