+++ BlazorServer/Pages/ClassConfigurationComponent.razor
@@ -1,8 +1,6 @@
-﻿@using System
-@using Core
-@using SharedLib
-@inject IBotController botController
+﻿@inject IBotController botController
 @inject IAddonReader addonReader
+
 @implements IDisposable
 
 <div class="container">
