+++ BlazorServer/Pages/EquipmentComponent.razor
@@ -1,8 +1,4 @@
-﻿@using SharedLib
-@using Core
-@using System.Xml
-
-<a href="@($"{WApi.ItemId}{Item.Entry}")" target="_blank" data-wowhead="item=@Item.Entry" class="small" style="text-decoration:none">
+﻿<a href="@($"{WApi.ItemId}{Item.Entry}")" target="_blank" data-wowhead="item=@Item.Entry" class="small" style="text-decoration:none">
     <img src="@imgLink" />
 </a>
