+++ BlazorServer/Pages/Gather.razor
@@ -1,6 +1,6 @@
 ﻿@page "/Gather"
-@using System.Threading.Tasks
-@inject Core.IBotController botController
+
+@inject IBotController botController
 @inject IJSRuntime JSRuntime
 
 @if (botController.ClassConfig?.Mode != Core.Mode.AttendedGather)
