+++ BlazorServer/Pages/GoalsComponent.razor
@@ -1,9 +1,7 @@
-﻿@using Core
-@inject Core.IBotController botController
-@inject Core.IAddonReader addonReader
-@implements IDisposable
+﻿@inject IBotController botController
+@inject IAddonReader addonReader
 
-@using SharedLib
+@implements IDisposable
 
 <div class="card" style="margin-top: 10px">
     <div class="card-header">
