+++ BlazorServer/Pages/InitButton.razor
@@ -1,5 +1,5 @@
-﻿@inject Core.IBotController botController
-@inject Core.IAddonReader addonReader
+﻿@inject IBotController botController
+@inject IAddonReader addonReader
 @inject AddonConfigurator addonConfigurator
 
 <button class="btn btn-sm btn-primary" @onclick="InitState">
