+++ BlazorServer/Pages/MiniMapComponent.razor
@@ -1,7 +1,4 @@
-﻿@using Core
-@using SharedLib
-@using Game
-@implements IDisposable
+﻿@implements IDisposable
 @inject IBotController botController
 @inject IAddonReader addonReader
