+++ BlazorServer/Pages/Money.razor
@@ -1,15 +1,13 @@
-﻿@if (Value.Length == 3)
-{
-    <span class="float-right" style="font-size:small">
-        @for (int i = 0; i < 2; i++)
+﻿<span class="float-right" style="font-size:small">
+    @for (int i = 0; i < 2; i++)
+    {
+        @if (Value[i] > 0)
         {
-            @if (Value[i] > 0)
-            {
-                @Value[i] <img height="12px" src="@images[i]" />
-            }
+            @Value[i] <img height="12px" src="@images[i]" />
         }
-    </span>
-}
+    }
+</span>
+
 @code {
     [Parameter]
     public int[] Value { get; set; } = new int[3] { 0, 0, 0 };
