+++ BlazorServer/Pages/PathSelectorComponent.razor
@@ -1,9 +1,9 @@
 ﻿@using MatBlazor
 @using System.IO
-@using SharedLib
-@using Core
+
 @inject IBotController botController
 @inject IAddonReader addonReader
+
 @implements IDisposable
 
 <div class="card @(Hide ? "hide" : "")" style="margin-top: 10px">
