+++ BlazorServer/Pages/ProfileSelectorComponent.razor
@@ -1,8 +1,8 @@
-﻿@using SharedLib
-@using System.IO
-@using Core
+﻿@using System.IO
+
 @inject IBotController botController
 @inject IAddonReader addonReader
+
 @implements IDisposable
 
 <div class="card @(Hide ? "hide" : "")" style="margin-top: 10px">
