+++ BlazorServer/Pages/RecordPath.razor
@@ -4,7 +4,7 @@
 @using System.IO
 @using System.Numerics
 @using SharedLib.Extensions
-@using Core
+
 @inject IAddonReader addonReader
 @inject IBotController botController
 @inject IJSRuntime JSRuntime
