+++ BlazorServer/Pages/RestartServer.razor
@@ -1,8 +1,8 @@
 ﻿@page "/RestartServer"
 
-@inject Microsoft.Extensions.Hosting.IHostApplicationLifetime lifetime
+@inject IHostApplicationLifetime lifetime
 @inject NavigationManager NavigationManager
-@inject Core.IBotController botController
+@inject IBotController botController
 
 <h2>Restarting the server - press reload when prompted...</h2>
