+++ BlazorServer/Pages/Screenshot.razor
@@ -1,5 +1,6 @@
 ﻿@page "/Screenshot"
-@inject Core.IBotController botController
+
+@inject IBotController botController
 
 <BotHeader />
 <ScreenshotComponent Stretch="true" />
