+++ BlazorServer/Pages/ScreenshotComponent.razor
@@ -1,5 +1,4 @@
-﻿@inject Core.IBotController botController
-@using Game;
+﻿@inject IBotController botController
 
 <div class="@(Stretch ? "w-100 h-auto m-0 p-0" : string.Empty)">
     <div class="row" @onclick="Toggle">
