+++ BlazorServer/Pages/Swag.razor
@@ -1,11 +1,9 @@
 ﻿@page "/Swag"
 
-@using Core
-@using SharedLib
-
 @inject IAddonReader addonReader
 @inject IBotController botController
 @inject IJSRuntime JSRuntime
+
 @implements IDisposable
 
 <BotHeader ShowActiveGoal="false" />
