+++ BlazorServer/Pages/_Host.cshtml
@@ -1,5 +1,4 @@
 ﻿@page "/"
-@using Microsoft.JSInterop
 @namespace BlazorServer.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @{
