+++ BlazorServer/Shared/MainLayout.razor
@@ -5,10 +5,6 @@
 </div>
 
 <div class="main">
-    @*<div class="top-row px-4">
-            <a href="https://docs.microsoft.com/aspnet/" target="_blank">About</a>
-        </div>*@
-
     <div class="content px-4">
         @Body
     </div>
