+++ BlazorServer/_Imports.razor
@@ -4,8 +4,12 @@
 @using Microsoft.AspNetCore.Components.Forms
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
+@using Microsoft.Extensions.Logging
+@using Microsoft.Extensions.Hosting
 @using Microsoft.JSInterop
 @using BlazorServer
 @using BlazorServer.Shared
-@using Core.Session
+@using SharedLib
+@using Core
+@using Game
 @using BlazorTable
