+++ Core/Addon/AddonDataProvider.cs
@@ -1,19 +1,20 @@
-﻿using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
+﻿using Game;
 using SharedLib;
-using Game;
+using System;
+using System.Collections.Generic;
+using System.Drawing;
 using System.Drawing.Imaging;
+using System.Linq;
 using System.Runtime.CompilerServices;
 
-namespace Core
+namespace Core.Addon
 {
-    public sealed class AddonDataProvider : IAddonDataProvider
+    public sealed class AddonDataProvider : IDisposable
     {
+        private readonly WowScreen wowScreen;
+
         private readonly DataFrame[] frames;
         private readonly int[] data;
-
-        private readonly WowScreen wowScreen;
 
         private readonly Color firstColor = Color.FromArgb(255, 0, 0, 0);
         private readonly Color lastColor = Color.FromArgb(255, 30, 132, 129);
@@ -22,23 +23,23 @@ namespace Core
         private readonly int bytesPerPixel;
 
         private readonly Rectangle bitmapRect;
+        private readonly Bitmap bitmap = null!;
 
         private Rectangle rect;
-        private readonly Bitmap bitmap = null!;
 
         public AddonDataProvider(WowScreen wowScreen, List<DataFrame> frames)
         {
             this.wowScreen = wowScreen;
 
             this.frames = frames.ToArray();
-            this.data = new int[this.frames.Length];
+            data = new int[this.frames.Length];
 
             rect.Width = frames.Last().point.X + 1;
             rect.Height = frames.Max(f => f.point.Y) + 1;
 
-            bytesPerPixel = Bitmap.GetPixelFormatSize(pixelFormat) / 8;
-            bitmap = new Bitmap(rect.Width, rect.Height, pixelFormat);
-            bitmapRect = new Rectangle(0, 0, bitmap.Width, bitmap.Height);
+            bytesPerPixel = Image.GetPixelFormatSize(pixelFormat) / 8;
+            bitmap = new(rect.Width, rect.Height, pixelFormat);
+            bitmapRect = new(0, 0, bitmap.Width, bitmap.Height);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -57,10 +58,10 @@ namespace Core
             {
                 BitmapData data = bitmap.LockBits(bitmapRect, ImageLockMode.ReadOnly, bitmap.PixelFormat);
 
-                byte* fLine = (byte*)data.Scan0 + (frames[0].point.Y * data.Stride);
+                byte* fLine = (byte*)data.Scan0 + frames[0].point.Y * data.Stride;
                 int fx = frames[0].point.X * bytesPerPixel;
 
-                byte* lLine = (byte*)data.Scan0 + (frames[^1].point.Y * data.Stride);
+                byte* lLine = (byte*)data.Scan0 + frames[^1].point.Y * data.Stride;
                 int lx = frames[^1].point.X * bytesPerPixel;
 
                 if (fLine[fx + 2] == firstColor.R && fLine[fx + 1] == firstColor.G && fLine[fx] == firstColor.B &&
@@ -68,10 +69,10 @@ namespace Core
                 {
                     for (int i = 0; i < frames.Length; i++)
                     {
-                        byte* line = (byte*)data.Scan0 + (frames[i].point.Y * data.Stride);
+                        byte* line = (byte*)data.Scan0 + frames[i].point.Y * data.Stride;
                         int x = frames[i].point.X * bytesPerPixel;
 
-                        this.data[frames[i].index] = (line[x + 2] * 65536) + (line[x + 1] * 256) + line[x];
+                        this.data[frames[i].index] = line[x + 2] * 65536 + line[x + 1] * 256 + line[x];
                     }
                 }
                 bitmap.UnlockBits(data);
@@ -90,3 +91,4 @@ namespace Core
         }
     }
 }
+
