--- Core/Addon/AddonDataProvider/IAddonDataProvider.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace Core
-{
-    public interface IAddonDataProvider : IDisposable
-    {
-        void Update();
-        int GetInt(int index);
-    }
-}
