+++ Core/Addon/AddonReader.cs
@@ -5,6 +5,7 @@ using Cyotek.Collections.Generic;
 using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Threading;
+using Core.Addon;
 
 namespace Core
 {
