+++ Core/StartupConfig/StartupConfigPathing.cs
